moved extract config binary to a package under container/go
